refactor: rename metadataSharing entity file
