+++ src/domain/entities/SharingSummary.ts
